requirements.txt
Django==5.2.7
reportlab==3.6.13
pdfrw==0.4
gunicorn==21.2.0
psycopg2-binary==2.9.9
Pillow==10.2.0
stripe==8.0.0
whitenoise==6.6.0  
python-dotenv==1.0.0
PyPDF2==3.0.1
whitenoise==6.7.0

